simplification total of all slugs

+++ server/apps/group/views.py
@@ -12,82 +12,63 @@ from django.contrib.sites.shortcuts import get_current_site
 from django.views.decorators.http import require_http_methods
 
 
-from .models import AdminRightsRequest, Group
-from apps.club.models import BDX
 from apps.sociallink.models import SocialLink
 from apps.event.models import BaseEvent
 from apps.post.models import Post
 
-from .forms import UpdateGroupForm, NamedMembershipAddGroup, NamedMembershipGroupFormset, SocialLinkGroupFormset, AdminRightsRequestForm
+from .forms import *
 
 from apps.utils.accessMixins import UserIsAdmin
+from apps.utils.slug import *
 
 
-class GroupSlugFonctions():
-    # group_type : nom de l'app correspondant au modèle demandé
-    # ------------ reproduit la chaîne devant le slug dans le modèle
-    # mini_slug : le slug du group, sans le type devant, reçu dans l'url
-    # ex:
-    # pour www.nantral-platform.fr/club/nantral-platform,
-    # on a group_type=club et mini_slug=nantral-platform
-    # la fonction ci-dessous renvoie alors slug="club--nantral-platform"
-
-    @property
-    def get_slug(self, **kwargs):
-        #group_type = self.kwargs.get('group_type')
-        app = resolve(self.request.path).app_name
-        mini_slug = self.kwargs.get("mini_slug")
-        # cas spécial du groupe club/bdx (mêmes urls)
-        if (app == "club"):
-            if BDX.objects.filter(slug='bdx--'+mini_slug):
-                return 'bdx--' + mini_slug
-            else:
-                return 'club--' + mini_slug
-        # autres groupes
-        else:
-            return app + '--' + mini_slug
 
 
-class BaseDetailGroupView(GroupSlugFonctions, DetailView):
-    '''Vue de détails d'un groupe.'''
+class BaseDetailGroupView(DetailView):
+    '''Vue de détails d'un groupe générique, sans protection.'''
     template_name = 'group/detail/detail.html'
 
     def get_object(self, **kwargs):
-        return Group.get_group_by_slug(self.get_slug)
+        app = resolve(self.request.path).app_name
+        slug = self.kwargs.get("slug")
+        return get_object_from_slug(app, slug)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         group = context['object']
-        context['admin_req_form'] = AdminRightsRequestForm()
+        #infos
+        context['sociallinks'] = SocialLink.objects.filter(
+            slug=group.full_slug)
+        context['events'] = BaseEvent.objects.filter(
+            group=group.full_slug, date__gte=date.today()).order_by('date')
+        context['posts'] = Post.objects.filter(
+            group=group.full_slug, publication_date__gte=date.today()-timedelta(days=10)
+        ).order_by('publication_date')
+        #members
         context['members'] = group.members.through.objects.filter(
             group=group).order_by('student__user__first_name')
-        context['form'] = NamedMembershipAddGroup(group)
-        context['sociallinks'] = SocialLink.objects.filter(
-            slug=group.slug)
         context['is_member'] = group.is_member(self.request.user)
-        if self.request.user.is_authenticated:
-            context['is_admin'] = group.is_admin(self.request.user)
-        else:
-            context['is_admin'] = False
-        context['events'] = BaseEvent.objects.filter(
-            group=group.slug, date__gte=date.today()).order_by('date')
-        context['posts'] = Post.objects.filter(
-            group=group.slug, publication_date__gte=date.today()-timedelta(days=10)
-        ).order_by('publication_date')
+        context['form'] = NamedMembershipAddGroup(group)
+        #admin
+        context['is_admin'] = group.is_admin(self.request.user)
+        context['admin_req_form'] = AdminRightsRequestForm()
         return context
 
 
 class DetailGroupView(LoginRequiredMixin, BaseDetailGroupView):
+    '''Vue de détail d'un groupe protégée.'''
     pass
 
 
-class AddToGroupView(LoginRequiredMixin, GroupSlugFonctions, FormView):
+class AddToGroupView(LoginRequiredMixin, FormView):
     '''Vue pour le bouton "Devenir Membre".'''
 
     raise_exception = True
 
     def get_group(self, **kwargs):
-        return Group.get_group_by_slug(self.get_slug)
+        app = resolve(self.request.path).app_name
+        slug = self.kwargs.get("slug")
+        return get_object_from_slug(app, slug)
 
     def form_valid(self, form):
         self.object = form.save(commit=False)
@@ -102,52 +83,57 @@ class AddToGroupView(LoginRequiredMixin, GroupSlugFonctions, FormView):
         return NamedMembershipAddGroup(group)
 
 
-class UpdateGroupView(UserIsAdmin, GroupSlugFonctions, TemplateView):
+class UpdateGroupView(UserIsAdmin, TemplateView):
     '''Vue pour modifier les infos générales sur un groupe.'''
 
     template_name = 'group/edit/update.html'
 
     def get_object(self, **kwargs):
-        return Group.get_group_by_slug(self.get_slug)
+        app = resolve(self.request.path).app_name
+        slug = self.kwargs.get("slug")
+        return get_object_from_slug(app, slug)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        self.object = self.get_object()
-        context['object'] = self.object
-        form_to_call = UpdateGroupForm(self.object)
-        if form_to_call:
-            context['form'] = form_to_call(instance=self.object)
+        context['object'] = self.get_object()
+        UpdateForm = UpdateGroupForm(context['object'])
+        if UpdateForm:
+            context['form'] = UpdateForm(instance=context['object'])
         return context
 
     def post(self, request, **kwargs):
         group = self.get_object()
-        form_to_call = UpdateGroupForm(group)
-        if form_to_call:
-            form = form_to_call(request.POST, request.FILES, instance=group)
-            form.save()
-        return redirect(group.group_type+':update', group.mini_slug)
+        UpdateForm = UpdateGroupForm(group)
+        if UpdateForm:
+            form = UpdateForm(request.POST, request.FILES, instance=group)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Informations modifiées !')
+            else:
+                messages.warning(request, form.errors)
+        return redirect(group.app+':update', group.slug)
 
 
-class UpdateGroupMembersView(UserIsAdmin, GroupSlugFonctions, TemplateView):
+class UpdateGroupMembersView(UserIsAdmin, TemplateView):
     '''Vue pour modifier les membres d'un groupe.'''
 
     template_name = 'group/edit/members_edit.html'
 
     def get_object(self, **kwargs):
-        return Group.get_group_by_slug(self.get_slug)
+        app = resolve(self.request.path).app_name
+        slug = self.kwargs.get("slug")
+        return get_object_from_slug(app, slug)
 
     def get_context_data(self, **kwargs):
         context = {}
         context['object'] = self.get_object()
         memberships = context['object'].members.through.objects.filter(
             group=context['object'])
-        membersForm = NamedMembershipGroupFormset(
-            context['object'])(queryset=memberships)
-        context['members'] = membersForm
+        MembersFormset = NamedMembershipGroupFormset(
+            context['object'])
+        if MembersFormset: 
+            context['members'] = MembersFormset(queryset=memberships)
         return context
-
-    # def get(self, request, mini_slug, group_type):
-        # return render(request, self.template_name, context=self.get_context_data(mini_slug=mini_slug))
 
     def post(self, request, **kwargs):
         group = self.get_object()
@@ -157,43 +143,40 @@ class UpdateGroupMembersView(UserIsAdmin, GroupSlugFonctions, TemplateView):
 @ require_http_methods(['POST'])
 @ login_required
 def edit_named_memberships(request, group):
-    form_to_call = NamedMembershipGroupFormset(group)
-    if form_to_call:
-        form = form_to_call(request.POST)
-    if form.is_valid():
-        members = form.save(commit=False)
-        for member in members:
-            member.group = group
-            member.save()
-        for member in form.deleted_objects:
-            member.delete()
-        messages.success(request, 'Membres modifies')
-        return redirect(group.group_type+':update-members', group.mini_slug)
-    else:
-        messages.warning(request, form.errors)
-        return redirect(group.group_type+':update-members', group.mini_slug)
+    MembersFormset = NamedMembershipGroupFormset(group)
+    if MembersFormset:
+        form = MembersFormset(request.POST)
+        if form.is_valid():
+            members = form.save(commit=False)
+            for member in members:
+                member.group = group
+                member.save()
+            for member in form.deleted_objects:
+                member.delete()
+            messages.success(request, 'Membres modifiés')
+        else:
+            messages.warning(request, form.errors)
+    return redirect(group.app+':update-members', group.slug)
 
 
-class UpdateGroupSocialLinksView(UserIsAdmin, GroupSlugFonctions, TemplateView):
+class UpdateGroupSocialLinksView(UserIsAdmin, TemplateView):
     '''Vue pour modifier les réseaux sociaux d'un groupe.'''
 
     template_name = 'group/edit/sociallinks_edit.html'
 
     def get_object(self, **kwargs):
-        return Group.get_group_by_slug(self.get_slug)
+        app = resolve(self.request.path).app_name
+        slug = self.kwargs.get("slug")
+        return get_object_from_slug(app, slug)
 
     def get_context_data(self, **kwargs):
         context = {}
         context['object'] = self.get_object()
-        slug = self.get_slug
-        sociallinks = SocialLink.objects.filter(slug=slug)
-        sociallinksForm = SocialLinkGroupFormset(queryset=sociallinks)
-        context['sociallinks'] = sociallinksForm
+        sociallinks = SocialLink.objects.filter(slug=context['object'].full_slug)
+        form = SocialLinkGroupFormset(queryset=sociallinks)
+        context['sociallinks'] = form
         return context
-
-    # def get(self, request, mini_slug, group_type):
-        # return render(request, self.template_name, context=self.get_context_data(mini_slug=mini_slug))
-
+    
     def post(self, request, **kwargs):
         group = self.get_object()
         return edit_sociallinks(request, group)
@@ -202,30 +185,29 @@ class UpdateGroupSocialLinksView(UserIsAdmin, GroupSlugFonctions, TemplateView):
 @ require_http_methods(['POST'])
 @ login_required
 def edit_sociallinks(request, group):
-    form_to_call = SocialLinkGroupFormset
-    if form_to_call:
-        form = form_to_call(request.POST)
+    form = SocialLinkGroupFormset(request.POST)
     if form.is_valid():
         sociallinks = form.save(commit=False)
         for sociallink in sociallinks:
-            sociallink.slug = group.slug
+            sociallink.slug = group.full_slug
             sociallink.save()
         for sociallink in form.deleted_objects:
             sociallink.delete()
         messages.success(request, 'Liens modifiés')
-        return redirect(group.group_type+':update-sociallinks', group.mini_slug)
     else:
         messages.warning(request, form.errors)
-        return redirect(group.group_type+':update-sociallinks', group.mini_slug)
+    return redirect(group.app+':update-sociallinks', group.slug)
 
 
 
-class RequestAdminRightsView(LoginRequiredMixin, GroupSlugFonctions, FormView):
+class RequestAdminRightsView(LoginRequiredMixin, FormView):
     raise_exception = True
     form_class = AdminRightsRequestForm
 
     def get_group(self, **kwargs):
-        return Group.get_group_by_slug(self.get_slug)
+        app = resolve(self.request.path).app_name
+        slug = self.kwargs.get("slug")
+        return get_object_from_slug(app, slug)
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -234,31 +216,33 @@ class RequestAdminRightsView(LoginRequiredMixin, GroupSlugFonctions, FormView):
 
     def form_valid(self, form):
         messages.success(
-            self.request, 'Votre demande a été enregistrée, on revient rapidement avec une réponse.')
+            self.request, 'Votre demande a bien été enregistrée ! Vous recevrez la réponse par mail.')
         object = form.save(commit=False)
         object.student = self.request.user.student
-        object.group = self.get_slug
+        object.group = self.get_group().full_slug
         object.save(domain=get_current_site(self.request).domain)
         return super().form_valid(form)
 
     def get_success_url(self) -> str:
         group = self.get_group()
-        return reverse(group.group_type+':detail', kwargs={'mini_slug': group.mini_slug})
+        return reverse(group.app+':detail', kwargs={'slug': group.slug})
 
 
-class AcceptAdminRequestView(UserIsAdmin, GroupSlugFonctions, View):
-    def get(self, request, mini_slug, id, group_type):
+class AcceptAdminRequestView(UserIsAdmin, View):
+    def get(self, request, slug, id):
         admin_req = AdminRightsRequest.objects.get(id=id)
         messages.success(
             request, message=f"Vous avez accepté la demande de {admin_req.student}")
         admin_req.accept()
-        return redirect(group_type+':detail', mini_slug)
+        app=reverse(request.path).app_name
+        return redirect(app+':detail', slug)
 
 
-class DenyAdminRequestView(UserIsAdmin, GroupSlugFonctions, View):
-    def get(self, request, mini_slug, id, group_type):
+class DenyAdminRequestView(UserIsAdmin, View):
+    def get(self, request, slug, id):
         admin_req = AdminRightsRequest.objects.get(id=id)
         messages.success(
             request, message=f"Vous avez refusé la demande de {admin_req.student}")
         admin_req.deny()
-        return redirect(group_type+':detail', mini_slug)
+        app=reverse(request.path).app_name
+        return redirect(app+':detail', slug)
